v1
From here on out it is a lot of fine tuning and adding the other
details. There are a few places I am not crazy about but some places
might not be bad once printed, sanded, filled in with wood filler and
primer fillered. Going to merge these changes into master and recreate
this v1 branch