v1
Created a duplicate of the helmet and moved it into the BAK collection
where other stages of the helmets stages were saved. Cleaned up the
other saves and applied the mirror and subdivision surface modifier and
made sure it would be an air tight print. There are a few more things to
do and a few places on the base helmet I am not crazy about however once
it is printed it will not be anything some sanding and wood filler will
not fix