v1
Started adding the side parts for the helmet and quickly realized the
back of the helmet had to level off for it to really work and match what
most of the references look like. I took and used a copy in the BAK
collection and fixed it. Should be good now on the helmet base